Add files via upload
The Login page with French as a globalization language on a test button working. I'm able to log into the database and get the correct error messages in the correct language depending on the boolean value triggered by the button press. The next step is to read the user's system and find out what region they are in so it can change to the correct language on startup.