Making some progress
I'm still working on the login with how it works right now there is no error for an invalid username as it's checking each user in the list. Not sure if this is the solution on going to stick with or not. I was able to display the customer information on the customer form Datagrid and will continue trying to learn how to get the customer insert to work.